for github

+++ ft_strlcpy.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        ::::::::            */
+/*   ft_strlcpy.c                                       :+:    :+:            */
+/*                                                     +:+                    */
+/*   By: fpurdom <fpurdom@student.codam.nl>           +#+                     */
+/*                                                   +#+                      */
+/*   Created: 2022/01/18 14:50:51 by fpurdom       #+#    #+#                 */
+/*   Updated: 2022/05/04 13:43:45 by fpurdom       ########   odam.nl         */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 #include <stddef.h>
 
@@ -5,7 +17,7 @@ size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 {
 	size_t	ret;
 
-	ret = ft_strlen(src);
+	ret = (size_t)ft_strlen(src);
 	if (dstsize == 0)
 		return (ret);
 	while (*src && dstsize - 1 > 0)
